requirements.txt
absl-py==0.11.0
aiodns==2.0.0
aiohttp==3.7.3
aiohttp-socks==0.5.5
asgiref==3.3.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.3.0
billiard==3.6.3.0
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.1
chardet
certifi==2020.12.5
cffi==1.14.4
click==7.1.2
coala-utils==0.5.1
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
dataclasses==0.6
decorator==4.4.2
distro==1.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
elasticsearch==7.10.1
erlastic==2.0.0
et-xmlfile==1.0.1
fake-useragent==0.1.11
filelock==3.0.12
flatbuffers==1.12
gast==0.3.3
geographiclib==1.50
geopy==2.1.0
grpcio==1.32.0
h5py==2.10.0
httplib2==0.18.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
importlib-metadata==3.4.0
jdcal==1.4.1
joblib==1.0.1
lxml==4.6.2
Markdown==3.3.3
moviepy==1.0.3
multidict==5.1.0
mysql-connector-python==8.0.22
numpy==1.19.5
oauthlib==3.1.0
openpyxl==3.0.5
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.2
params-flow==0.8.2
pickle5==0.0.11
Pillow==8.1.0
proglog==0.1.9
protobuf==3.14.0
py-params==0.10.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==3.1.1
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-slugify==3.0.2
python-socks==1.2.0
pytz==2021.1
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
sacremoses==0.0.43
schedule==0.6.0
scipy==1.6.0
six==1.15.0
slugify==0.0.1
SmoothAPI==1.0.3
soupsieve==2.1
sqlparse==0.4.1
tabula-py==2.2.0
termcolor==1.1.0
text-unidecode==1.2
tqdm==4.56.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
Werkzeug==1.0.1
wget==3.2
wrapt==1.12.1
xlrd==2.0.1
yarl==1.6.3
zipp==3.4.0


